requirements.txt
Bio==1.7.1
biopython==1.85
biotite==1.1.0
fair_esm==2.0.0
matplotlib==3.10.0
modelgenerator==0.1.1.post5
numpy==2.2.2
pytorch_lightning==2.5.0.post0
scikit_learn==1.6.1
torch==2.6.0
tqdm==4.67.1
hydra-core==1.3.2
git+https://github.com/genbio-ai/openfold.git@c4aa2fd0d920c06d3fd80b177284a22573528442
git+https://github.com/NVIDIA/dllogger.git@0540a43971f4a8a16693a9de9de73c1072020769